refactor: reorganize dashboard components and update imports
