+++ src/app/dashboard/layout.tsx
@@ -3,7 +3,7 @@ import Footer from "@/components/layout/footer/Footer";
 import { SidebarInset, SidebarProvider } from "@/components/ui/sidebar";
 import Header from "@/components/layout/header/Header";
 import { Button } from "@/components/ui/button";
-import { DateRangePicker } from "@/components/ui/dashboard/Datepicker";
+import { DateRangePicker } from "@/components/modules/dashboard/Datepicker";
 
 const layout = ({ children }: { children: React.ReactNode }) => {
   return (
@@ -12,17 +12,7 @@ const layout = ({ children }: { children: React.ReactNode }) => {
       <SidebarInset>
         <Header />
         <div className="bg-muted/50 min-h-screen">
-          {/* Este contenedor crece dinámicamente con el contenido */}
-          <div className="flex-1 space-y-4 p-4 pt-6 md:p-8">
-            <div className="flex items-center justify-between">
-              <h2 className="text-3xl font-bold tracking-tight">Dashboard</h2>
-              <div className="flex items-center space-x-2">
-                <DateRangePicker />
-                <Button>Download</Button>
-              </div>
-            </div>
-            {children}
-          </div>
+          <div className="flex-1 space-y-4 p-4 pt-6 md:p-8">{children}</div>
         </div>
         <Footer />
       </SidebarInset>
